Hangman
Hangman game where your lives are accounted for and each wrong guess displays a part of the Hangman. Duplicate letter entries from user are tracked and notified (no loose of life for this). Incorrect letters are notified (lose one life).
Game ends when remaining lives reaches 0 or when user guesses all the correct letter or when user guesses the correct word.