Caesar Cipher
Program where user is prompted to choose between encoding or decoding a word. The user will provide inputs such as the word to process and how many numbers to shift each letter. The shift is performed using a List of letters in the whole alphabet, i.e.. 'a' shifted by 3 will result in the letter 'd'. The user is prompted until they choose to exit the program.